Rewrite the logic used to handle Redis commands with clustering.
The reason for this change is to support the upcoming redis-cluster since it
has a few different behaviours compared to the default client-side cluster
provided by Predis, for example redis-cluster will not support key tags or
certain operations that are currently available with out client-side
implementation.

This re-implementation is actually slightly slower than the previous one but
we still need to optimize things.