Add initial support for Redis cluster (Redis v3.0).
This implementation is capable of properly handling `ASK` and `MOVED`
replies returned by Redis when one of the nodes asks the client for
a temporary or a permanent redirect for the slot to a different node.

Performances looks almost on par with our client-side sharding solution
and it looks already relatively stable.